MACROS Parsing and Layers Parsing complete!

+++ src/Parser/Vector.java
@@ -2,15 +2,15 @@ package Parser;
 
 
 public class Vector {
-    float x, y, z;
+    double x, y, z;
 
-    Vector (float x , float y , float z) {
+    Vector (double x , double y , double z) {
         this.x = x ;
         this.y = y ;
         this.z = z ;
     }
     
-    Vector (float x , float y ) {
+    Vector (double x , double y ) {
         this.x = x ;
         this.y = y ;
     }
